pyproject.toml
[project]
name = "nohow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[openai]>=1.2.0",
    "mirascope[openai]>=1.25.7",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.45",
    "textual>=6.11.0",
]

[dependency-groups]
dev = [
    "textual-dev>=1.8.0",
]
